feat(graphic): integrate shadow into the deferred pipeline

+++ src/plugin/default-pipeline/src/component/GPUDirectionalLight.hpp
@@ -15,7 +15,7 @@ struct GPUDirectionalLight {
     Id buffer{};
     Id bindGroupData{};
     wgpu::TextureView shadowTextureView;
-    size_t shadowTextureIndex;
+    uint32_t shadowTextureIndex;
 
     void Update(const Object::Component::DirectionalLight &light, const Object::Component::Transform &transform)
     {
